requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.1
schedule==1.2.0
loguru==0.7.2
numpy==1.25.2
scipy==1.11.3
pydantic==2.4.2
sqlalchemy==2.0.23
alembic==1.12.0
psycopg2-binary==2.9.9
dvc==3.30.1
pytest
apscheduler
colorama